[Dev][PhotoDataSource] Load all albumIds first, and use albumId to load photos (can't assume that albumIds are continuous)

+++ app/src/main/java/com/xdd/cloudinteraction2019/data/model/Photo.kt
@@ -4,8 +4,8 @@ import androidx.recyclerview.widget.DiffUtil.ItemCallback
 import com.google.gson.annotations.SerializedName
 
 data class Photo(
-    @SerializedName("albumId") val albumId: Int,
-    @SerializedName("id") val id: Int,
+    @SerializedName("albumId") val albumId: Long,
+    @SerializedName("id") val id: Long,
     @SerializedName("title") val title: String,
     @SerializedName("url") val url: String,
     @SerializedName("thumbnailUrl") val thumbnailUrl: String
